GOVSI-523 - Validate the ID token hint in the Logout request
- If the ID token hint exists then check if it exists under any client_session_id in the users session. If it doesn't then thrown an exception. The clientSessionId and the id token hint will not always correlate in the logout request as the clientSessionId will always be the last client they used whereas the idTokenHint will come from whatever RP they are logging out from.
- Validate that the clientSessionID in the cookies exists in the users session and if it doesn't then throw an exception.
- Put in a placeholder to validate the signature of the ID token. Once we have a key which we are not just generating on the fly then we can implement this.

+++ serverless/lambda/src/main/java/uk/gov/di/lambdas/LogoutHandler.java
@@ -4,12 +4,16 @@ import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent;
+import uk.gov.di.entity.ClientSession;
 import uk.gov.di.entity.Session;
+import uk.gov.di.helpers.CookieHelper;
 import uk.gov.di.services.ConfigurationService;
 import uk.gov.di.services.SessionService;
 
 import java.util.Map;
 import java.util.Optional;
+
+import static java.lang.String.format;
 
 public class LogoutHandler
         implements RequestHandler<APIGatewayProxyRequestEvent, APIGatewayProxyResponseEvent> {
@@ -40,14 +44,54 @@ public class LogoutHandler
     private APIGatewayProxyResponseEvent processLogoutRequest(
             Session session, APIGatewayProxyRequestEvent input) {
         Map<String, String> queryStringParameters = input.getQueryStringParameters();
+        Optional<CookieHelper.SessionCookieIds> sessionCookieIds =
+                CookieHelper.parseSessionCookie(input.getHeaders());
 
+        boolean sessionContainsClientSessionID =
+                session.getClientSessions()
+                        .containsKey(sessionCookieIds.get().getClientSessionId());
+
+        if (!sessionContainsClientSessionID) {
+            throw new RuntimeException(
+                    format(
+                            "Client Session ID does not exist in Session: %s",
+                            session.getSessionId()));
+        }
         if (!queryStringParameters.containsKey("id_token_hint")
                 || queryStringParameters.get("id_token_hint").isBlank()) {
             sessionService.deleteSessionFromRedis(session.getSessionId());
             return generateDefaultLogoutResponse();
         }
+        if (!doesIDTokenExistInSession(queryStringParameters.get("id_token_hint"), session)) {
+            throw new RuntimeException(
+                    format("ID Token does not exist for Session: %s", session.getSessionId()));
+        }
+        if (!validateIDTokenSignature(
+                queryStringParameters.get("id_token_hint"), session.getSessionId())) {
+            throw new RuntimeException(
+                    format(
+                            "Unable to validate ID token signature for Session: %s",
+                            session.getSessionId()));
+        }
+
         sessionService.deleteSessionFromRedis(session.getSessionId());
         return generateDefaultLogoutResponse();
+    }
+
+    private boolean validateIDTokenSignature(String idTokenHint, String sessionID) {
+        return true;
+    }
+
+    private boolean doesIDTokenExistInSession(String idTokenHint, Session session) {
+        for (Map.Entry<String, ClientSession> t : session.getClientSessions().entrySet()) {
+            boolean idTokenHintExists =
+                    t.getValue().getIdTokenHint() != null
+                            && t.getValue().getIdTokenHint().equals(idTokenHint);
+            if (idTokenHintExists) {
+                return true;
+            }
+        }
+        return false;
     }
 
     private APIGatewayProxyResponseEvent generateDefaultLogoutResponse() {
